Automatically update the Web Starter Kit latest release download URL
Grunt 'build' will trigger the 'wsk-version' task, which in turn uses
['latest-release'](https://www.npmjs.org/package/latest-release) to
determine the name, tag and zip-archive download URL of the latest
release of google/web-starter-kit on Github.

The results of this are written to src/site/_config-wsk-version.yml.

Jekyll.destination.config has been expanded to pick this config yml up
load in the variables when building the site.

src/site/_en/starter-kit/index.html uses the {{site.wsk-zip-url}}
variable to point it's download button the to latest version.

See discussion in:
- google/web-starter-kit#414
- google/WebFundamentals#547