Make the appengine and devsite builds be (almost) the same.
The code for wrapping pages in <html>...</html>, etc, is now in the GAE script,
so the built files themselves don't need to have it. This means that we only
need one type of build.

There is still one difference, as devsite builds link to the samples on Github
rather than having them locally.