headless chrome: use execFile()
exec() spawns a shell, which means that unsanitised user input must not
be passed. The example suggests doing exactly that. Presumably people
will use Lighthouse, but nevertheless I don't think such a potentially
dangerous approach should be shown as an example.

This commit changes the example to use execFile(). This does not spawn
a shell, avoiding the security risks of exec().