Fixed some broken images
There doesn't appear to be any JavaScript on the page that lazy
loads images with the `lazyload` class. And given that there are
many images on the page, and only two use this class, it may be
because this article was moved, and two `img` elements were not
updated. Also, the `small` class on the same two images doesn't
seem to have any style associated with it, and it is not used on any
other image on the page. Therefore, the `lazyload` and `small` classes
were removed from the two images, and `src` is now used instead of
`data-src` to specify the image URL.