Update to Node.js v10 (#6769)
This also moves the Node.js version number to an .nvmrc file, so that
people using `nvm` can just enter `nvm use` to install and switch to
the expected Node.js version for this project. Travis CI uses the
`.nvmrc` file as well, so this is the single source of truth.