emscripten-npm: improve --host & --target flags to configure
When using standard autoconf configure scripts, the flags are:
* --build: The system where the project is being compiled.
* --host: The system where the compiled project will execute.
* --target: The system which the compiled project will generate for.

Thus, when cross-compiling, the standard flag to pass is --host.
Most projects don't even support --target as that tends to be for
projects like compilers.  The libvpx project is unfortunate in that
it provides an autoconf-like configure script but with the wrong
use of --host & --target options.

Clarify the article to focus on --host and mention --target as a
non-standard/uncommon thing which it is.