Improve instructions.
Addresses #7665 
I think the problem here is a general one. In may cases, a link such as this one is tangential to the page it contains, creating the expectation the link does not need to be followed to understand the page's content. 

In this case, the link is central to what the page is trying to accomplish. This change make it clear that the link *must* be followed to start the task described on the page.