Modified triggerPushMsg() in Web Push article (#7735)
* Modified triggerPushMsg() in Web Push article

Made the `if` statement consistent with a paragraph in the article
which states that in case of 404 and 410, the subscription has expired
or is no longer valid and so we need to remove the subscription from
the database.
Also, in the `else` statement we are now throwing `err` rather than
letting `undefined` be returned. We do this because in case of other
errors, such as 400 (Invalid request) or 429 (Too many requests),
the promise returned by triggerPushMsg() should just reject.
Finally, added a paragraph which explains that in case of an error other
than 404 and 410, we `throw err` to let the promise reject.

* Changed console.error to console.log

Since we're just logging that a subscription has expired, there is no need
to use console.error.