Update KV Storage syntax to use default export
As pointed out in this [pull request in the spec's GitHub repo for KV storage](https://github.com/WICG/kv-storage/pull/74). The KV storage api now uses a default export instead of a named export. This can be tested in Chrome Canary with experimental web features turned on and is also present in the current spec.