$ no longer an alias for document.querySelector
It could be in the past, the function is an alias for `document.querySelector`, but if the statement  $: "this function is an alias for the document.querySelector() function" is passed around, it might cause some confusion...

```// not exactly an alias
document.querySelector === $   // false```   

The real situation is, when called with one argument, this function (`$`) calls `document.querySelector()`
But when called with $(sel, node), calls `node.querySelector(sel)`. I think in the past, it dosen't accept a second argument, and it was an alias for `document.querySelector()` but not any more.