Update Device Orientation sample code.
Although this sample is not linked by any pages since #6714, Googling for
"device orientation demo" still shows the googlesamples.github.io version of
this code as one of the first results. As such, fix the "deviceorientation"
event handler so that it applies rotations in a more correct and natural
way:
* Invert across the 3 axes rather than just 2, and do so in the order
  mandated by the Device Orientation spec.
* Invert the signs of the alpha and gamma rotations to account for
  differences between the CSS and Device Orientation coordinate systems.

Starting with Chrome 89, some of the presets in the DevTools Sensors tab
will cause the HTML5 logo not to show up because of the angles
involved (this is also the case even without this commit), but slight
changes in the rotations will cause it to be shown again.

This change was triggered by https://crbug.com/1137281 and the behavior of
the sample while I tried to investigate the DevTools behavior.